+++ Source/Krypton Components/TestForm/DataGridViewDemo.cs
@@ -191,6 +191,19 @@ public partial class DataGridViewDemo : KryptonForm
         kdgvMain.HideOuterBorders = !kchkShowGridLines.Checked;
     }
 
+    private void kchkCornerRounding_CheckedChanged(object sender, EventArgs e) => ApplyCornerRoundingFromDemo();
+
+    private void knudCornerRoundingRadius_ValueChanged(object sender, EventArgs e) => ApplyCornerRoundingFromDemo();
+
+    private void ApplyCornerRoundingFromDemo()
+    {
+        knudCornerRoundingRadius.Enabled = kchkCornerRounding.Checked;
+
+        float rounding = kchkCornerRounding.Checked ? (float)knudCornerRoundingRadius.Value : -1f;
+        kdgvMain.StateNormal.Border.Rounding = rounding;
+        kdgvMain.StateDisabled.Border.Rounding = rounding;
+    }
+
     private void kcmbAutoSizeColumnsMode_SelectedIndexChanged(object sender, EventArgs e)
     {
         if (kcmbAutoSizeColumnsMode.SelectedItem is string name)